[feat] use warn to inform user error might be ok
When you are provisioning a cluster kubeform will attempt to check to see
if the cluster already exists -- this results in an error if the cluster
is being made new.  The error it emits is ignorable in this case.

Switch the log level to warn to stop scaring people and insert text allowing
the operator to make the appropriate call given the scenario