(chore) remove dependency injection (#5)
Injecting kubeform and the debug outputter from the cli invoker was very
confusing and makes it difficult to wire up other tools to use the built-in
commands, in addition, if you were to include other parts of the system you'd
get no log output at all. This switches over to use pino and it's pretty printer
and export JSON logs, unless you're invoking from the CLI.

Required flags should have sane defaults -- the onboarding documentation recommends
writing out the auth information in a specific location, so we should use that.
Users who have done something different may still override it, but this allows
the rest of us to be lazy.