WIP: Add amazon provider, refactor
Start adding AWS as a provider. Move to using classes that extend from a
BaseClass instead of the dependency injection that was currently implemented.

This provides a clearer layout of what needs to be implemented for a provider;
the BaseProvider class has four methods that throw if not implemented. All
other methods are internal. The base class should replace the metadata file
in each provider -- there might be supplimental pieces of provider-specific data
but those should be imported and attached/overridden in the provider-specific
subclass of the base object

Instead of relying on closed over variables being bound to functions and re-exported
those variables are now captured by the class and stored as instance properties.

Refactor how zones are kept so that they can be re-used by the both the inquire
method and the individual providers.